Update session04.rst
I'm not sure about the change on lines 770-771, but it struck me as odd.  And then my suspicion was confirmed on slide 49 (lines 975-6), and the file wsgi_1.py.


Line 1022 says:
You may also want to consider using ``print environ`` in your application so
you can see the dictionary.

But the wsgi_1.py file already has a pprint environ statement in it.